Fix: make crm/services.h externally usable again - temporary hack
